rag/requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.5
cbor==1.0.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
coloredlogs==15.0.1
datasets==4.0.0
dill==0.3.8
distro==1.9.0
fastapi==0.116.1
fastembed==0.7.2
filelock==3.19.1
FlagEmbedding==1.3.5
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
gast==0.6.0
google-pasta==0.2.0
groq==0.31.0
grpcio==1.74.0
h11==0.16.0
h2==4.3.0
h5py==3.14.0
hf-xet==1.1.9
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
ijson==3.4.0
inscriptis==2.6.0
ir_datasets==0.5.11
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.11.3
langchain==0.3.27
langchain-core==0.3.75
langchain-groq==0.3.7
langchain-ollama==0.3.7
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.10
langsmith==0.4.20
libclang==18.1.1
loguru==0.7.3
lxml==6.0.1
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
mmh3==5.2.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
namex==0.1.0
networkx==3.5
numpy==2.1.3
ollama==0.5.3
onnxruntime==1.22.1
opt_einsum==3.4.0
optree==0.17.0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
peft==0.17.1
pillow==11.3.0
portalocker==3.2.0
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
py_rust_stemmers==0.1.5
pyarrow==21.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.15.1
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.4
trec-car-tools==2.6
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
unlzw3==0.2.3
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
zlib-state==0.1.9
zstandard==0.24.0
